+++ WeatherBotApi.DatabaseApp/Core/DatabaseApp.Application/Users/Commands/CreateUser/CreateUserCommand.cs
@@ -3,9 +3,9 @@ using MediatR;
 
 namespace DatabaseApp.Application.Users.Commands.CreateUser;
 
-public class CreateUserCommand : IRequest<Result<int>>
+public class CreateUserCommand : IRequest<Result<long>>
 {
-    public required int TelegramId { get; init; }
+    public required long TelegramId { get; init; }
     public required string Username { get; init; }
     public required string MobileNumber { get; init; }
     public required DateTime RegisteredAt { get; init; }
