+++ WeatherBotApi.DatabaseApp/Core/DatabaseApp.Domain/Models/User.cs
@@ -3,7 +3,7 @@
 public class User : IEntity
 {
     public int Id { get; init; }
-    public int TelegramId { get; init; }
+    public long TelegramId { get; init; }
     public required UserMetadata Metadata { get; init; }
     public DateTime RegisteredAt { get; init; }
 }
